components/pytorch-hello-world-backend/requirements.txt
Flask==2.0.2
Flask-CORS==3.0.10
Pillow==8.3.2
